:building_construction: Split mainactivity into fragments
