+++ app/src/main/java/com/kyhsgeekcode/disassembler/AnalysisResultFragment.kt
@@ -1,0 +1,17 @@
+package com.kyhsgeekcode.disassembler
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import kotlinx.android.synthetic.main.fragment_analysis_result.*
+
+class AnalysisResultFragment: Fragment(){
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?) =
+            inflater.inflate(R.layout.fragment_analysis_result, container, false)!!
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        imageViewCount.setOnClickListener(this)
+    }
+}
