+++ app/src/main/java/com/kyhsgeekcode/disassembler/HexFragment.kt
@@ -1,0 +1,72 @@
+package com.kyhsgeekcode.disassembler
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.MotionEvent
+import android.view.View
+import android.view.ViewGroup
+import android.widget.AdapterView
+import androidx.fragment.app.Fragment
+import kotlinx.android.synthetic.main.fragment_hexview.*
+//TODO: Add a cusom HEX view
+class HexFragment : Fragment() {
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?) =
+            inflater.inflate(R.layout.fragment_analysis_result, container, false)!!
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+
+        mainGridViewHex.setOnTouchListener { v: View, event: MotionEvent ->
+            if (touchSource == null) touchSource = v
+            if (v === touchSource) {
+                mainGridViewAscii.dispatchTouchEvent(event)
+                if (event.action == MotionEvent.ACTION_UP) {
+                    clickSource = v
+                    touchSource = null
+                }
+            }
+            false
+        }
+        mainGridViewHex.onItemClickListener = AdapterView.OnItemClickListener { parent: AdapterView<*>, view: View?, position: Int, id: Long ->
+            if (parent === clickSource) { // Do something with the ListView was clicked
+            }
+        } /*
+		gvHex.setOnScrollListener(new OnScrollListener() {
+				@Override
+				public void onScroll(AbsListView view, int firstVisibleItem, int visibleItemCount, int totalItemCount) {
+					if(view == clickSource)
+						gvAscii.setSelectionFromTop(firstVisibleItem, view.getChildAt(0).getTop() + offset);
+				}
+
+				@Override
+				public void onScrollStateChanged(AbsListView view, int scrollState) {}
+			});*/
+        mainGridViewAscii.setOnTouchListener { v, event ->
+            if (touchSource == null) touchSource = v
+            if (v === touchSource) {
+                mainGridViewHex.dispatchTouchEvent(event)
+                if (event.action == MotionEvent.ACTION_UP) {
+                    clickSource = v
+                    touchSource = null
+                }
+            }
+            false
+        }
+        mainGridViewAscii.setOnItemClickListener { parent, view, position, id ->
+            if (parent === clickSource) { // Do something with the ListView was clicked
+            }
+        }
+        /*
+		gvAscii.setOnScrollListener(new OnScrollListener() {
+				@Override
+				public void onScroll(AbsListView view, int firstVisibleItem, int visibleItemCount, int totalItemCount) {
+					if(view == clickSource)
+						gvHex.setSelectionFromTop(firstVisibleItem, view.getChildAt(0).getTop()/ * + offset);
+				}
+
+				@Override
+				public void onScrollStateChanged(AbsListView view, int scrollState) {}
+			});
+			*/
+    }
+}
