+++ app/src/main/java/com/kyhsgeekcode/disassembler/LogFragment.kt
@@ -1,0 +1,19 @@
+package com.kyhsgeekcode.disassembler
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import nl.lxtreme.binutils.elf.MachineType
+import java.util.*
+
+class LogFragment : Fragment(){
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?) =
+            inflater.inflate(R.layout.fragment_project_overview, container, false)!!
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        refreshlog.setOnClickListener(this)
+        loglistView.adapter = LogAdapter().also { logAdapter = it }
+    }
+}
