+++ app/src/main/java/com/kyhsgeekcode/disassembler/ViewPagerAdapter.kt
@@ -1,0 +1,28 @@
+package com.kyhsgeekcode.disassembler
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentStatePagerAdapter
+
+class ViewPagerAdapter(supportFragmentManager: FragmentManager) : FragmentStatePagerAdapter(supportFragmentManager,BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
+
+    private val mFragmentList = ArrayList<Fragment>()
+    private val mFragmentTitleList = ArrayList<String>()
+
+    override fun getItem(position: Int): Fragment {
+        return mFragmentList[position]
+    }
+
+    override fun getCount(): Int {
+        return mFragmentList.size
+    }
+
+    override fun getPageTitle(position: Int): CharSequence? {
+        return mFragmentTitleList[position]
+    }
+
+    fun addFragment(fragment: Fragment, title: String) {
+        mFragmentList.add(fragment)
+        mFragmentTitleList.add(title)
+    }
+}
