+++ app/src/main/res/values/strings.xml
@@ -94,4 +94,7 @@
     <string name="foundstrings">Strings</string>
     <string name="analysis">Analysis</string>
 
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
 </resources>
